pyproject.toml
[project]
name = "backup-config-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "python-decouple>=3.8",
    "supabase>=2.18.1",
]
